Add files via upload
We will implement two different networks to perform automatic rating (0 or 1) of a movie given the text of its review. We will use the imdb (internet movie database) dataset.

The reviews are already available in the form of indexes that point to a word dictionary: each word is already encoded as an index in the dictionary.